v1.2/requirements.txt
# Core Dependencies
asyncio
PyYAML>=6.0

# LLM Interface (Example: OpenAI)
openai>=1.0.0
# Add other LLM providers
anthropic>=0.20.0
google-generativeai>=0.5.0 # Add Google Gemini

# Memory System (Example: Redis)
redis>=4.0.0

# Browser Agent (Example: Playwright)
playwright>=1.30.0

# Optional Web UI/API (Example: Streamlit/FastAPI)
streamlit>=1.20.0
# fastapi>=0.100.0
# uvicorn[standard]>=0.20.0

# Optional Agent Frameworks (Uncomment as needed)
# langchain>=0.1.0
# crewai>=0.20.0
# autogen>=0.2.0

# Optional Vector DBs (Uncomment as needed)
# chromadb>=0.4.0
chromadb>=0.4.24
# weaviate-client>=3.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.20.0 # For testing async code
# Add mocking libraries if needed
# pytest-mock
# unittest.mock (built-in)